backend/requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
marshmallow==3.26.2
psycopg[binary]==3.2.6
PyJWT==2.10.1
python-dotenv==1.0.1
gunicorn==21.2.0